requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
boto3==1.15.3
botocore==1.18.18
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.12
django-auto-logout==0.5.0
django-braces==1.15.0
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-etc==1.3.1
django-filter==22.1
django-forms-dynamic==1.0.0
django-htmx==1.9.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-mptt==0.13.4
django-phonenumber-field==6.1.0
django-star-ratings==0.9.2
django-storages==1.13.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
folium==0.12.1.post1
fonttools==4.33.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
Jinja2==3.1.1
jmespath==0.10.0
kiwisolver==1.4.3
lxml==4.9.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.56
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.27.1
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
swapper==1.3.0
tablib==3.2.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.25.11
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
